+++ Java/Sort an Array/README.md
@@ -1,2 +1,1 @@
-
 [Utkarsh Nagar](https://github.com/Utkarshn10)
